+++ Common/src/main/java/no/fjordkraft/im/if320/models/DistributionDetails.java
@@ -16,7 +16,7 @@ public class DistributionDetails {
     @XmlElement(name = "Type", required = true)
     protected String type;
     @XmlElement(name = "Amount")
-    protected float amount;
+    protected double amount;
 
     public String getType() {
         return type;
@@ -26,11 +26,11 @@ public class DistributionDetails {
         this.type = type;
     }
 
-    public float getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
 }
